Add another GStreamer based encoder for FLV videos.
There's now 3 encoders to chose from (with reason of existance)
1) GIF (default)
Good for screencasts where changes don't happen constantly and are
local, easy to export to the web.
2) Theora
Good for screencasts with frequent changes where detail is not that
important, easy to export to the web.
3) Flash Screen video
Lossless encoding (if you want to postprocess in Pitivi), same area of
usage as GIF, but harder to export to the web (needs a Flash environment
like Youtube).