Add an intermediae serialization format and tools to handle it
The intermediate format is intended to quickly dump data to a file for
later processing. It does not aim at reducing file size, being
compatible to any other format.
The intended use is as a quick storage in /tmp that ensures recording
can continue even when the CPU cannot keep up with encoding the image
stream. This rarely happens on modern CPUs, but can often happen on
older or embedded devices.

Additions:
- byzanzserialize.[ch], code to (de)serialize internal data to/from a
  stream.
- ByzanzEncoderByzanz, an encoder that writes out that intermediate
  format to a file. Extension is .byzanz.
- byzanz-playback, a simple application that reads files in .byzanz
  format and encodes them using the Byzanz encoders.

This allows to encode a file for later processing, possibly even
multiple times. The main intended use however is benchmarking the
performance-critical parts of the Byzanz pipeline.